Performance improvements
* Add sleeps into the update loop so that the application sleeps until
  the next opcode or update
* Improve audio to remove deprecated rust feature
* Improve video device to only redraw if pixels have changed
* Add roms for testing (mostly pulled from another chip8 emulator git
  repo)

TODO: Improve the key polling opcode as this can cause the application
to be unresponsive until the key is pressed

+++ src/chip8/mod.rs
@@ -25,8 +25,8 @@ struct Args {
     pixelsize: usize,
 
     // the clock speed on the "cpu" in MHz
-    #[clap(short, long, default_value = "1.0")]
-    clockspeed: f32,
+    #[clap(short, long, default_value = "1000")]
+    clockspeed: u32,
 }
 
 pub fn start() {
